+++ src/WiFiConnectionManager.h
@@ -52,19 +52,19 @@ class WiFiConnectionManager : public ConnectionManager {
 
     void changeConnectionState(NetworkConnectionState _newState);
 
-    const int CHECK_INTERVAL_IDLE = 100;
-    const int CHECK_INTERVAL_INIT = 100;
-    const int CHECK_INTERVAL_CONNECTING = 500;
-    const int CHECK_INTERVAL_CONNECTED = 10000;
-    const int CHECK_INTERVAL_RETRYING = 5000;
-    const int CHECK_INTERVAL_DISCONNECTED = 1000;
-    const int CHECK_INTERVAL_ERROR = 500;
+    const unsigned long CHECK_INTERVAL_IDLE = 100;
+    const unsigned long CHECK_INTERVAL_INIT = 100;
+    const unsigned long CHECK_INTERVAL_CONNECTING = 500;
+    const unsigned long CHECK_INTERVAL_CONNECTED = 10000;
+    const unsigned long CHECK_INTERVAL_RETRYING = 5000;
+    const unsigned long CHECK_INTERVAL_DISCONNECTED = 1000;
+    const unsigned long CHECK_INTERVAL_ERROR = 500;
 
     const char *ssid, *pass;
     unsigned long lastConnectionTickTime;
 
     WiFiClient wifiClient;
-    int connectionTickTimeInterval;
+    unsigned long connectionTickTimeInterval;
 };
 
 #endif /* #ifdef BOARD_HAS_WIFI */
